==> 4_predict_winner/1_feature_engineering.ipynb <==
import pandas as pd
import re
pd.set_option('display.max_columns', None)

import warnings
warnings.filterwarnings("ignore")

### Objective

> in order to predict the outcome based on fighting style + others I need to mimic the stats that I can input for upcoming fights -- that info is available 

1. -- split columns like x of y into x_attempted and x_successful -- these are:  
- f1_sig_str + f2_sig_str
- f1_td + f2_td


2. -- get fight time in seconds   
-length_in_seconds
3. 
| Statistic                                                   | Calculation                                                           |
|-------------------------------------------------------------|-----------------------------------------------------------------------|
| SLpM - Significant Strikes Landed per Minute                | fx_slmp = (fx_sig_str_success / length_in_seconds) x 60              |
| Str. Acc. - Significant Striking Accuracy                   | fx_str_acc = (fx_sig_str_success / f1_sig_str_attempted) x 100      |
| SApM - Significant Strikes Absorbed per Minute             | fx_sapm = fy_slmp                                                    |
| Str. Def. - Significant Strike Defence                      | fx_str_def = 100 - fy_str_acc                                         |
| TD Avg. - Average Takedowns Landed per 15 minutes           | fx_td_avg = (fx_td_success / length_in_seconds) x 60 * 15           |
| TD Acc. - Takedown Accuracy                                 | fx_td_acc = fx_td_%                                                   |
| TD Def. - Takedown Defense                                  | fx_td_def = 100 - fy_td_%                                             |
| Sub. Avg. - Average Submissions Attempted per 15 minutes    | fx_sub_avg = (fx_sub_att / length_in_seconds) x 60                   |


# Load the dataset
data = pd.read_csv('../3_cluster_fight_style/data_with_win_percentages.csv')
data.shape

**1. -- split columns like x of y into x_attempted and x_successful -- these are:**

# Function to split '2 out of 3' into two separate columns
def split_success_attempted(df, column_name):
    # Split the column into a DataFrame with two columns
    split_data = df[column_name].str.strip().str.split(expand=True)
    
    # Assign the split data to new columns
    df[f'{column_name}_success'] = pd.to_numeric(split_data[0], errors='coerce')
    df[f'{column_name}_attempted'] = pd.to_numeric(split_data[2], errors='coerce')

columns_to_split = ['f1_sig_str', 'f2_sig_str', 'f1_td', 'f2_td']

for column in columns_to_split:
    split_success_attempted(data, column)

# Drop the original columns if needed
data.drop(columns=columns_to_split, inplace=True)

# Check the transformed DataFrame
display(data)

data['gender'].value_counts()

**2. -- get fight time in seconds** 

def time_to_seconds(time_str):
    # Split the time
    minutes, seconds = map(int, time_str.split(':'))
    
    # Convert hours and minutes to seconds
    total_seconds = minutes * 60 + seconds
    
    return total_seconds

data['end_time_seconds'] = data['end_time'].apply(time_to_seconds)
# get time in seconds of last round + (rounds - one + 300) for full rounds
data['length_in_seconds'] = data['end_time_seconds'] + (data['round'] -1 ) * 300
data.drop(columns='length_in_seconds')
data[['length_in_seconds', 'end_time', 'round']]

data.isna().sum()[data.isna().sum() > 0]

import dis
import pandas as pd

# Calculate statistics for fighter 1 (f1)
fight_stats = pd.DataFrame()

fight_stats['method'] = data['method']
fight_stats['referee'] = data['referee']
fight_stats['weightclass'] = data['weightclass']
fight_stats['winner'] = 'f1'
fight_stats['date'] = data['fight_date']

fight_stats['f1'] = data['f1']
fight_stats['f2'] = data['f2']

fight_stats['f1_slmp'] = (data['f1_sig_str_success'] / data['length_in_seconds']) * 60
fight_stats['f2_slmp'] = (data['f2_sig_str_success'] / data['length_in_seconds']) * 60

fight_stats['f1_str_acc_%'] = (data['f1_sig_str_success'] / data['f1_sig_str_attempted']) * 100
fight_stats['f2_str_acc_%'] = (data['f2_sig_str_success'] / data['f2_sig_str_attempted']) * 100

fight_stats['f1_sapm'] = fight_stats['f2_slmp']
fight_stats['f2_sapm'] = fight_stats['f1_slmp']

fight_stats['f1_str_def_%'] = 100 - fight_stats['f2_str_acc_%']
fight_stats['f2_str_def_%'] = 100 - fight_stats['f1_str_acc_%']

fight_stats['f1_td_avg'] = (data['f1_td_success'] / data['length_in_seconds']) * 60 * 15
fight_stats['f2_td_avg'] = (data['f2_td_success'] / data['length_in_seconds']) * 60 * 15

data['f1_td_%'] = pd.to_numeric(data['f1_td_%'].str.rstrip('%'), errors='coerce')
data['f2_td_%'] = pd.to_numeric(data['f2_td_%'].str.rstrip('%'), errors='coerce')

fight_stats['f1_td_acc_%'] = data['f1_td_%'] 
fight_stats['f2_td_acc_%'] = data['f2_td_%'] 

fight_stats['f1_td_def_%'] = 100 - data['f2_td_%']  
fight_stats['f2_td_def_%'] = 100 - data['f1_td_%']  

fight_stats['f1_sub_avg'] = (data['f1_sub_att'] / data['length_in_seconds']) * 60
fight_stats['f2_sub_avg'] = (data['f2_sub_att'] / data['length_in_seconds']) * 60

fight_stats['f1_height'] = data['f1_height'] 
fight_stats['f2_height'] = data['f2_height'] 

fight_stats['f1_weight'] = data['f1_weight'] 
fight_stats['f2_weight'] = data['f2_weight'] 

fight_stats['f1_reach'] = data['f1_reach'] 
fight_stats['f2_reach'] = data['f2_reach'] 

fight_stats['f1_stance'] = data['f1_stance'] 
fight_stats['f2_stance'] = data['f2_stance'] 

fight_stats = fight_stats.round(2)

fight_stats.isna().sum()[fight_stats.isna().sum() > 0]

columns_with_null = ['f1_str_acc_%', 'f2_str_acc_%', 'f1_str_def_%', 'f2_str_def_%']
rows_with_null = fight_stats[fight_stats[columns_with_null].isnull().any(axis=1)]
display(rows_with_null.head(2))
columns_with_null = ['f1_str_acc_%', 'f2_str_acc_%', 'f1_str_def_%', 'f2_str_def_%']
rows_with_null_index = fight_stats[fight_stats[columns_with_null].isnull().any(axis=1)].index
rows_with_null_data = data.loc[rows_with_null_index]
display(rows_with_null_data.head(2))

# nan value bc dividing by 0 
fight_stats.fillna('0', inplace=True)
fight_stats.isna().sum().sum()

fight_stats.head(3)

fight_stats.to_csv('../feature_engineered_fight.csv', index=False)
