==> 4_predict_winner/4_predict_302.ipynb <==
import pickle
import pandas as pd

# Load the model and preprocessor
with open('gradient_boosting_pipeline.pkl', 'rb') as file:
    loaded_pipeline = pickle.load(file)

# Example of making predictions on new data
Makhachev_Poirier = pd.DataFrame({
    'f1_stance': ['Orthodox'],
    'f2_stance': ['Orthodox'],
    'weightclass': ['lightweight'],
    'f1_slmp': [2.46],
    'f2_slmp': [5.45],
    'f1_str_acc_%': [60.0],
    'f2_str_acc_%': [51.0],
    'f1_sapm': [1.27],
    'f2_sapm': [4.36],
    'f1_str_def_%': [61.0],
    'f1_td_avg': [3.17],
    'f2_td_avg': [1.33],
    'f1_td_acc_%': [60.0],
    'f2_td_acc_%': [36.0],
    'f1_td_def_%': [90.0],
    'f2_td_def_%': [63.0],
    'f1_sub_avg': [1.0],
    'f2_sub_avg': [1.4],
    'f1_reach': [70],
    'f2_reach': [72]
}) 

# Predict the winner
predicted_winner = loaded_pipeline.predict(Makhachev_Poirier)
print("Predicted Winner:", predicted_winner[0])

Strickland_Costa = pd.DataFrame({
    'f1_stance': ['Orthodox'],
    'f2_stance': ['Orthodox'],
    'weightclass': ['middleweight'],
    'f1_slmp': [5.91],
    'f2_slmp': [6.20],
    'f1_str_acc_%': [41.0],
    'f2_str_acc_%': [58.0],
    'f1_sapm': [4.33],
    'f2_sapm': [6.38],
    'f1_str_def_%': [62.0],
    'f1_td_avg': [0.85],
    'f2_td_avg': [0.45],
    'f1_td_acc_%': [64.0],
    'f2_td_acc_%': [75.0],
    'f1_td_def_%': [77.0],
    'f2_td_def_%': [80.0],
    'f1_sub_avg': [0.2],
    'f2_sub_avg': [0.0],
    'f1_reach': [76],
    'f2_reach': [72]
}) 

# Predict the winner
predicted_winner = loaded_pipeline.predict(Strickland_Costa)
print("Predicted Winner:", predicted_winner[0])

Holland_Oleksiejczuk = pd.DataFrame({
    'f1_stance': ['Orthodox'],
    'f2_stance': ['Unorthodox'],
    'weightclass': ['middleweight'],
    'f1_slmp': [4.24],
    'f2_slmp': [5.06],
    'f1_str_acc_%': [49.0],
    'f2_str_acc_%': [50.0],
    'f1_sapm': [3.19],
    'f2_sapm': [4.26],
    'f1_str_def_%': [50.0],
    'f1_td_avg': [0.85],
    'f2_td_avg': [0.75],
    'f1_td_acc_%': [38.0],
    'f2_td_acc_%': [44.0],
    'f1_td_def_%': [54.0],
    'f2_td_def_%': [48.0],
    'f1_sub_avg': [0.6],
    'f2_sub_avg': [0.0],
    'f1_reach': [81],
    'f2_reach': [74]
}) 

# Predict the winner
predicted_winner = loaded_pipeline.predict(Holland_Oleksiejczuk)
print("Predicted Winner:", predicted_winner[0])

Price_Morono = pd.DataFrame({
    'f1_stance': ['Orthodox'],
    'f2_stance': ['Orthodox'],
    'weightclass': ['welterweight'],
    'f1_slmp': [5.46],
    'f2_slmp': [4.87],
    'f1_str_acc_%': [42.0],
    'f2_str_acc_%': [40.0],
    'f1_sapm': [5.60],
    'f2_sapm': [4.16],
    'f1_str_def_%': [48.0],
    'f1_td_avg': [0.70],
    'f2_td_avg': [0.32],
    'f1_td_acc_%': [20.0],
    'f2_td_acc_%': [22.0],
    'f1_td_def_%': [64.0],
    'f2_td_def_%': [55.0],
    'f1_sub_avg': [0.7],
    'f2_sub_avg': [0.4],
    'f1_reach': [76],
    'f2_reach': [72]
}) 

# Predict the winner
predicted_winner = loaded_pipeline.predict(Price_Morono)
print("Predicted Winner:", predicted_winner[0])

Brown_Santos = pd.DataFrame({
    'f1_stance': ['Orthodox'],
    'f2_stance': ['Orthodox'],
    'weightclass': ['welterweight'],
    'f1_slmp': [4.37],
    'f2_slmp': [4.52],
    'f1_str_acc_%': [46.0],
    'f2_str_acc_%': [42.0],
    'f1_sapm': [3.33],
    'f2_sapm': [3.33],
    'f1_str_def_%': [54.0],
    'f1_td_avg': [0.78],
    'f2_td_avg': [0.43],
    'f1_td_acc_%': [39.0],
    'f2_td_acc_%': [13.0],
    'f1_td_def_%': [74.0],
    'f2_td_def_%': [68.0],
    'f1_sub_avg': [0.5],
    'f2_sub_avg': [0.4],
    'f1_reach': [78],
    'f2_reach': [73]
}) 

# Predict the winner
predicted_winner = loaded_pipeline.predict(Brown_Santos)
print("Predicted Winner:", predicted_winner[0])
